Extract test-utils
