+++ packages/playground/server-components/tests/worker-e2e.test.ts
@@ -1,4 +1,4 @@
-import {port} from './worker-e2e.serve';
+import {port} from '../../test-utils/worker-e2e.serve';
 import testCases from './e2e-test-cases';
 
 describe('Worker', () => {
