+++ packages/playground/test-utils/load-prod-env.js
@@ -1,7 +1,7 @@
 const {loadEnv} = require('vite');
 
-module.exports.loadProdEnv = async function () {
-  const env = await loadEnv('production', __dirname, '');
+module.exports.loadProdEnv = async function (cwd = __dirname) {
+  const env = await loadEnv('production', cwd, '');
   Object.keys(env).forEach((key) => {
     if (['VITE_', 'PUBLIC_'].some((prefix) => key.startsWith(prefix))) {
       delete env[key];
