+++ packages/playground/test-utils/node-prod-e2e.serve.ts
@@ -1,22 +1,21 @@
 // This is automtically detected by scripts/jest-e2e-setup-test.ts and will replace
 // the default e2e test serve behavior
 
-import {resolve} from 'path';
 import {execSync} from 'child_process';
+import {createServer} from './start-node';
 
-export const port = 9528;
+export const port = 9428;
 
 export async function serve(root: string, isProd: boolean) {
   // we build first, regardless of whether it's prod/build mode
   // because Vite doesn't support the concept of a "webworker server"
   execSync(
-    'yarn build:client --logLevel silent && yarn build:worker --logLevel silent',
+    'yarn build:client --logLevel silent && yarn build:server --logLevel silent',
     {cwd: root}
   );
 
   // @ts-ignore
-  const {createServer} = await import(resolve(root, 'start-worker.js'));
-  const {app} = await createServer({root, isProd});
+  const {app} = await createServer({cwd: root});
 
   return new Promise((resolve, reject) => {
     try {
